Add patch to write use dbl prec flag to btScalar.h
If the CMake option USE_DOUBLE_PRECISION is true, btScalar.h will be configured
during build to contain the lines
```
  #ifndef BT_USE_DOUBLE_PRECISION
  #define BT_USE_DOUBLE_PRECISION
  #endif
```
Otherwise, btScalar.h will not contain those lines. This lets us sidestep the
issue of the pods macros not reading the preprocessor flags from pkg-config
files, and still not have to define the preprocessor flag everywhere that
includes the bullet headers.